theme update in widgets

+++ src/hikogui/theme/theme_model_tests.cpp
@@ -8,14 +8,14 @@
 TEST(theme_model, pixel_length)
 {
     struct delegate_type {
-        [[nodiscard]] std::pair<hi::theme_state, float> state_and_scale() const noexcept {
-            return {hi::theme_state::enabled, -2.0f};
+        [[nodiscard]] std::pair<hi::theme_state, int> state_and_scale() const noexcept {
+            return {hi::theme_state::enabled, -2};
         }
     };
 
     auto &m = hi::theme_model_by_key("pixel-length-test");
-    m[hi::theme_state::enabled].width = hi::pixels{42.0};
+    m[hi::theme_state::enabled].width = hi::pixels{42};
 
     auto d = delegate_type{};
-    ASSERT_FLOAT_EQ(hi::theme<"pixel-length-test">.width(&d), 42.0f);
+    ASSERT_FLOAT_EQ(hi::theme<"pixel-length-test">.width(&d), 42);
 }
